# modules/fbs-core/api/api-docs.yml
openapi: 3.0.0
info:
  title: Feedbacksystem
  description: Feedbacksystem API
  version: 1.0.0
  contact:
    email: Andrej.Sajenko@mni.thm.de
  license:
    name: CC BY-NC-SA 4.0
    url: https://creativecommons.org/licenses/by-nc-sa/4.0/
...
security:
  - JWT: [ ]
paths:
...
  /users:
    get:
      operationId: getUsers

      tags:
        - User
      summary: Get user list
      description: ""
      responses:
        "200":
          description: OK
          headers:
            Authorization:
              description: Bearer
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
    post:
      operationId: createUser

      tags:
        - User
      summary: Create a new user
      description: ""
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        description: User object to create
        required: true
      responses:
        "201":
          description: Created
          headers:
            Authorization:
              description: Bearer
              schema:
                type: string
        "400":
          description: Invalid Format
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "409":
          description: Already Existing
...
  "/users/{uid}/courses/{cid}/tasks/{tid}/submissions":
    get:
      operationId: getCourseTaskSubmissions

      tags:
        - Submission
      summary: Get all submissions for a task
      parameters:
        - name: uid
          in: path
          description: user id
          required: true
          schema:
            type: integer
        - name: cid
          in: path
          description: course id
          required: true
          schema:
            type: integer
        - name: tid
          in: path
          description: task id
          required: true
          schema:
            type: integer
        - in: query
          name: passed
          description: Filters only the passed submissions
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          headers:
            Authorization:
              description: Bearer
              schema:
                type: string
          content:
            "*/*":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
    post:
      operationId: submitCourseTaskSolution

      tags:
        - Submission
      summary: Submit a solution for a task
      parameters:
        - name: uid
          in: path
          description: user id
          required: true
          schema:
            type: integer
        - name: cid
          in: path
          description: course id
          required: true
          schema:
            type: integer
        - name: tid
          in: path
          description: task id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          "*/*":
            schema:
              format: binary
        description: The solution for the task
        required: true
      responses:
        "201":
          description: Created
          headers:
            Authorization:
              description: Bearer
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
  "/users/{uid}/courses/{cid}/tasks/{tid}/submissions/{sid}/content":
    get:
      operationId: getCourseTaskSubmissionsContent
      tags:
        - Submission
      summary: Get the submission content for a task
      parameters:
        - name: uid
          in: path
          description: user id
          required: true
          schema:
            type: integer
        - name: cid
          in: path
          description: course id
          required: true
          schema:
            type: integer
        - name: tid
          in: path
          description: task id
          required: true
          schema:
            type: integer
        - name: sid
          in: path
          description: submission id
          required: true
          schema:
            type: integer
        - in: query
          name: passed
          description: Filters only the passed submissions
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          headers:
            Authorization:
              description: Bearer
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: tar
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
  "/users/courses/{cid}/tasks/submissions/content":
    get:
      operationId: solutionsOfCourse
      tags:
        - Submission
      summary: Get the submission content for a course
      parameters:
        - name: cid
          in: path
          description: course id
          required: true
          schema:
            type: integer
        - in: query
          name: passed
          description: Filters only the passed submissions
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          headers:
            Authorization:
              description: Bearer
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: tar
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
  "/users/courses/{cid}/tasks/{tid}/submissions/content":
    get:
      operationId: solutionsOfTask
      tags:
        - Submission
      summary: Get the submission content for a task
      parameters:
        - name: cid
          in: path
          description: course id
          required: true
          schema:
            type: integer
        - name: tid
          in: path
          description: task id
          required: true
          schema:
            type: integer
        - in: query
          name: passed
          description: Filters only the passed submissions
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          headers:
            Authorization:
              description: Bearer
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: tar
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
  "/users/{uid}/courses/{cid}/tasks/{tid}/submissions/{sid}/subresults":
    get:
      operationId: getCourseTaskSubmissionSubresults

      tags:
        - Submission
      summary: Get all subresults for a submission
      parameters:
        - name: uid
          in: path
          description: user id
          required: true
          schema:
            type: integer
        - name: cid
          in: path
          description: course id
          required: true
          schema:
            type: integer
        - name: tid
          in: path
          description: task id
          required: true
          schema:
            type: integer
        - name: sid
          in: path
          description: submission id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          headers:
            Authorization:
              description: Bearer
              schema:
                type: string
          content:
            "*/*":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubTaskResult"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
  "/users/{uid}/courses/{cid}/tasks/{tid}/submissions/{sid}":
    put:
      operationId: restartCourseTaskSubmission

      tags:
        - Submission
      summary: Restart the submission process
      parameters:
        - name: uid
          in: path
          description: user id
          required: true
          schema:
            type: integer
        - name: cid
          in: path
          description: course id
          required: true
          schema:
            type: integer
        - name: tid
          in: path
          description: task id
          required: true
          schema:
            type: integer
        - name: sid
          in: path
          description: ID of submission
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          headers:
            Authorization:
              description: Bearer
              schema:
                type: string
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
    get:
      operationId: getCourseTaskSubmission

      tags:
        - Submission
      summary: Get the current submission state
      parameters:
        - name: uid
          in: path
          description: user id
          required: true
          schema:
            type: integer
        - name: cid
          in: path
          description: course id
          required: true
          schema:
            type: integer
        - name: tid
          in: path
          description: task id
          required: true
          schema:
            type: integer
        - name: sid
          in: path
          description: ID of submission
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          headers:
            Authorization:
              description: Bearer
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
...
servers:
  - url: https://feedback.mni.thm.de/api/v1
components:
...
  securitySchemes:
    JWT:
      type: apiKey
      in: header
      name: Authorization
      description: Bearer
  schemas:
    User:
      type: object
      required:
        - username
        - prename
        - surname
        - postname
        - email
        - password
      properties:
        id:
          type: integer
        prename:
          type: string
        surname:
          type: string
        email:
          type: string
        password:
          type: string
          description: Only for creation
        username:
          type: string
        alias:
          type: string
        globalRole:
          type: string
          description: "Either, ADMIN, MODERATOR, or USER. Default is USER"
...
    Submission:
      type: object
      required:
        - submissionTime
        - done
        - id
        - results
      properties:
        submissionTime:
          type: integer
          description: Unix time as long
        done:
          type: boolean
          description: True if all checkers have checked the submissions
        id:
          type: integer
          description: The submission id
        results:
          type: array
          description: The check results of each configured checker, may be empty
          items:
            $ref: "#/components/schemas/CheckResult"
    CheckResult:
      type: object
      required:
        - exitCode
        - resultText
        - checkerType
        - configurationId
      properties:
        exitCode:
          type: integer
          description: Exit code of the check result 0 is okay
        resultText:
          type: string
          description: The output the check
        checkerType:
          type: string
          description: The type of the checker that checked the submission
        configurationId:
          type: integer
          description: The configuration id of the checker configuration
        extInfo:
          type: object
          nullable: true
          additionalProperties: true
          description: Extended information of the check
...
    SubTaskResult:
      type: object
      properties:
        name:
          type: string
        points:
          type: number
        maxPoints:
          type: number
...